refactor: normalize the lack of `PropTypes`
the react v19 has dropped support for prop types in favour of typescript.
this little fix aims to normalize this lack of support while keeping it
in a javascript environment deliberately. for the time being, this is
how it will be used. in the future, i might consider migrating to type-
script, should this library prove useful enough.